refactor: rename UserController to UserRestController
